chore: organize safe files & dirs
