+++ apps/dapp/package.json
@@ -19,7 +19,7 @@
     "version:release": "standard-version",
     "version:prerelease": "standard-version --prerelease beta",
     "build-with-stats": "tsc && BUILD_STATS=true vite build",
-    "codegen:safe-api-v1": "openapi --input ./safeOpenApi/apiV1.json --output ./safeOpenApi/client"
+    "codegen:safe-api-v1": "openapi --input ./src/safe/open-api/apiV1.json --output ./src/safe/open-api/client"
   },
   "dependencies": {
     "@balancer-labs/sdk": "^1.0.4",
