+++ apps/dapp/tsconfig.json
@@ -20,9 +20,9 @@
       "data/*": ["./data/*"],
       "types/*": ["./types/*"],
       "constants/*": ["./constants/*"],
-      "safeApi/*": ["../safeOpenApi/*"],
+      "safe/*": ["./safe/*"],
     },
   },
-  "include": ["**/*.ts", "**/*.tsx","**/*.json", "api/ping.js", "api/geoblock.js", "safeOpenApi/*"],
+  "include": ["**/*.ts", "**/*.tsx","**/*.json", "api/ping.js", "api/geoblock.js"],
   "exclude": ["node_modules", "./dist/"]
 }
